Implement deferred service registration
This is the first part of handling deferred service providers. Yet, now if a service provider is deferred, the application will automatically listen for it "when" events and register when dispatched.

Services that are offered via the "provides" are marked for a later point. The IoC logic now needs to be adapted.