Fix command registration from config
Well... the idea was nice to use the kernel for registration, while the service provider remained deferred. BUT - this caused dual creation of the Artisan (Console Application) instance, because this provider was booted, before the Kernel's artisan property was set. This was a serious "chicken or egg" dilemma.

Anyway, this solution is actually much more elegant, in that we no longer care about the kernel and go strait to the console application.