Change client, make it formally container aware
The problem with the prev. revisions was that the container "magically" had to be known. This was a problem for the builder that just magically relied on it being available in the constructor - which normally it is, but developer is not able to set it before pipes are in use.