Simplify circuit breaker, remove conditional trip mechanism
The idea was to only trip the circuit breaker on certain types of exceptions, but this logic make the design much more complex than it has to be. Developers should implement their own way of dealing with any / all kinds of exceptions, whilst the circuit breaker ultimately should treat all kinds of exceptions as failures.

#9