Extract processing logic into own component
I realized that if a very large amount of items, e.g. database records, are passed into the Summation instance, a reference would be kept which could cause a large amount of memory to be filled.
By extracting processing logic, item reference is kept to a minimum.
Thus, the summation collection is as "clean" as can be expected.

#27