Change key components, use a Repo. & specification pattern instead
Processing rules can now implement the "Determinable" interface, which then tells whether or not a rule can be applied for given item.
The Repository is able to return a collection of rules that match given item (those that can process item).
Lastly, the item processor's process method is no longer static, which allows a different type of implementation.

#27