Fix unable to handle force-delete events
To my big suprise, Laravel actually only serialises a model's class path, identifier and database connection. Upon un-serialisation of an event, the model is queried from the database. I suppose this is good enough when dealing with queued events, so that it's always the latest version from a database that is used. Yet, it does pose a slight problem when dealing with actual force deleted models.
Thus, the changes here just populate an empty model with an id, so the rest of the code does not break.