Change Mime-Type dto, cache information
In case that the same mime-type object is accessed multiple times, there could be an overhead on the sampler / File Info instance. This is especially true if a resource has been used. So, this should ensure that the object only queries the sampler once - unless the cached information is cleared.