Redesign ETag generator
After some time working with this, it just felt cumbersome to have multiple "profiles" for generating weak / strong ETags. Therefore, I decided to allow developers to determine this when using the given generator, on the spot. Furthermore, the Eloquent Model specific generator is no more. Instead, a simple trait will ensure that a model can generate a weak / strong ETag representation of it. This also allows each model to customize what attributes to use for an ETag.