Change resource context, allow arbitrary data to be set
The resource now also holds an actual record, file or some kind of content that the request is intended to yield. Furthermore, arbitrary data can no be associated with the resource. This is going to be very useful for complex setting or changing states that other parts of an application must react upon.