Fix missing distinguished delimiter
For multipart response body, a last delimiter (boundary separator) was needed, acc. to RFC 2046. This also entailed resolving an issue with removal of newline characters in the multipart response (test util).

Lastly, the test document is also updated to contain a last newline character for the sake of testing these changes.