Ignore exception handler
Sadly, no good solution can be made in this major version without introducing a breaking change. Either the interface is respected and a Response instance is returned, an exception is thrown, or exit is invoked (behavior change is breaking).
For now, this is ignored. But, in next major version, this defect must be addressed.

#153