Fix incorrect assert of log-file location
It appears that APP_ENV might has bleed over from a previous test, which is why the log file location would had been "athenaeum" and not "laravel". However, something has change or been fixed in 3rd party components, which means that now the test is actually performing as it should. There is nothing that specifies that APP_ENV should be set to "testing", for these tests - so the default values should be used here.